requirements.txt
reportlab
requests
pandas
matplotlib
pytz
psycopg2-binary
prophet
plotly
django
djangorestframework
python-decouple
dj-database-url